Enlever dupplication code filter
